LayoutTests/fast/filesystem/workers/simple-persistent-sync.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
worker = startWorker('../resources/simple-persistent-sync.js');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
