LayoutTests/fast/filesystem/workers/simple-persistent-sync-expected.txt

This tests requesting PERSISTENT FileSystemSync in Worker.

Starting worker: ../resources/simple-persistent-sync.js
Description: requestFileSystemSync PERSISTENT test.
Got error from worker: Uncaught ReferenceError: requestFileSystemSync is not defined

TEST COMPLETE
